Implemented MethodQ

+++ IMethodQ.cs
@@ -13,7 +13,7 @@ namespace MethodQ
         /// <param name="action">The <see cref="Action"/> to schedule.</param>
         /// <param name="executionUtc">The <see cref="DateTime"/> in UTC of when the <paramref name="action"/> should be invoked.</param>
         /// <returns>The scheduled method call's unique identifier (needed for cancellation).</returns>
-        ulong Schedule(Action action, DateTime executionUtc);
+        int Schedule(Action action, DateTime executionUtc);
 
         /// <summary>
         /// Schedules the specified <see cref="Action"/> to be executed every <paramref name="interval"/> until manual cancellation.
@@ -21,13 +21,13 @@ namespace MethodQ
         /// <param name="action">The <see cref="Action"/> to schedule.</param>
         /// <param name="interval">Repetition interval <see cref="TimeSpan"/>.</param>
         /// <returns>The scheduled method call's unique identifier (needed for cancellation).</returns>
-        ulong Schedule(Action action, TimeSpan interval);
+        int Schedule(Action action, TimeSpan interval);
 
         /// <summary>
         /// Cancels the specified method call (removes it from the list of scheduled calls). Irreversible!
         /// </summary>
         /// <param name="id">The scheduled action's identifier.</param>
         /// <returns>Whether the cancellation was successful or not (e.g. can't double-cancel or cancel an inexistent call).</returns>
-        bool Cancel(ulong id);
+        bool Cancel(int id);
     }
 }
